+++ Code/Engine/Core/Platform/Android/WindowDecl_Platform.h
@@ -1,0 +1,5 @@
+
+struct ANativeWindow;
+using ezWindowHandle = ANativeWindow*;
+using ezWindowInternalHandle = ezWindowHandle;
+#define INVALID_WINDOW_HANDLE_VALUE nullptr
