+++ Code/Engine/Core/Platform/Android/Window_Platform.h
@@ -1,0 +1,5 @@
+
+// can't use a 'using' here, because that can't be forward declared
+class EZ_CORE_DLL ezWindow : public ezWindowPlatformShared
+{
+};
