+++ Code/Engine/Core/Platform/NoImpl/WindowDecl_NoImpl.h
@@ -1,0 +1,4 @@
+
+using ezWindowHandle = void*;
+using ezWindowInternalHandle = ezWindowHandle;
+#define INVALID_WINDOW_HANDLE_VALUE nullptr
