+++ Code/Engine/Core/Platform/OSX/WindowDecl_Platform.h
@@ -1,0 +1,2 @@
+
+#include <Core/Platform/NoImpl/WindowDecl_NoImpl.h>
