+++ Code/Engine/Core/Platform/UWP/WindowDecl_Platform.h
@@ -1,0 +1,4 @@
+
+using ezWindowHandle = IUnknown*;
+using ezWindowInternalHandle = ezWindowHandle;
+#define INVALID_WINDOW_HANDLE_VALUE nullptr
