+++ Code/Engine/Core/System/Window.h
@@ -11,125 +11,12 @@ class ezOpenDdlWriter;
 class ezOpenDdlReader;
 class ezOpenDdlReaderElement;
 
-
-// Include the proper Input implementation to use
-#if EZ_ENABLED(EZ_SUPPORTS_GLFW)
-#  include <Core/Platform/GLFW/InputDevice_GLFW.h>
-#else
-#  include <InputDevice_Platform.h>
-#endif
-
 // Currently the following scenarios are possible
 // - Windows native implementation, using HWND
 // - GLFW on windows, using GLFWWindow* internally and HWND to pass windows around
 // - GLFW / XCB on linux. Runtime uses GLFWWindow*. Editor uses xcb-window. Tagged union is passed around as window handle.
 
-#if EZ_ENABLED(EZ_SUPPORTS_GLFW)
-
-extern "C"
-{
-  typedef struct GLFWwindow GLFWwindow;
-}
-
-#  if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-#    include <Foundation/Platform/Win/Utils/MinWindows.h>
-using ezWindowHandle = ezMinWindows::HWND;
-using ezWindowInternalHandle = GLFWwindow*;
-#    define INVALID_WINDOW_HANDLE_VALUE (ezWindowHandle)(0)
-#    define INVALID_INTERNAL_WINDOW_HANDLE_VALUE nullptr
-#  elif EZ_ENABLED(EZ_PLATFORM_LINUX)
-
-extern "C"
-{
-  typedef struct xcb_connection_t xcb_connection_t;
-}
-
-struct ezXcbWindowHandle
-{
-  xcb_connection_t* m_pConnection;
-  ezUInt32 m_Window;
-};
-
-struct ezWindowHandle
-{
-  enum class Type
-  {
-    Invalid = 0,
-    GLFW = 1, // Used by the runtime
-    XCB = 2   // Used by the editor
-  };
-
-  Type type;
-  union
-  {
-    GLFWwindow* glfwWindow;
-    ezXcbWindowHandle xcbWindow;
-  };
-
-  bool operator==(ezWindowHandle& rhs)
-  {
-    if (type != rhs.type)
-      return false;
-
-    if (type == Type::GLFW)
-    {
-      return glfwWindow == rhs.glfwWindow;
-    }
-    else
-    {
-      // We don't compare the connection because we only want to know if we reference the same window.
-      return xcbWindow.m_Window == rhs.xcbWindow.m_Window;
-    }
-  }
-};
-
-using ezWindowInternalHandle = ezWindowHandle;
-#    define INVALID_WINDOW_HANDLE_VALUE \
-      ezWindowHandle {}
-#  else
-using ezWindowHandle = GLFWwindow*;
-using ezWindowInternalHandle = GLFWwindow*;
-#    define INVALID_WINDOW_HANDLE_VALUE (GLFWwindow*)(0)
-#  endif
-
-#elif EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-
-#  include <Foundation/Platform/Win/Utils/MinWindows.h>
-using ezWindowHandle = ezMinWindows::HWND;
-using ezWindowInternalHandle = ezWindowHandle;
-#  define INVALID_WINDOW_HANDLE_VALUE (ezWindowHandle)(0)
-
-#elif EZ_ENABLED(EZ_PLATFORM_WINDOWS_UWP)
-
-using ezWindowHandle = IUnknown*;
-using ezWindowInternalHandle = ezWindowHandle;
-#  define INVALID_WINDOW_HANDLE_VALUE nullptr
-
-#elif EZ_ENABLED(EZ_PLATFORM_ANDROID)
-
-struct ANativeWindow;
-using ezWindowHandle = ANativeWindow*;
-using ezWindowInternalHandle = ezWindowHandle;
-#  define INVALID_WINDOW_HANDLE_VALUE nullptr
-
-#elif EZ_ENABLED(EZ_PLATFORM_WEB)
-
-struct DummyWindowHandle;
-using ezWindowHandle = DummyWindowHandle*;
-using ezWindowInternalHandle = ezWindowHandle;
-#  define INVALID_WINDOW_HANDLE_VALUE nullptr
-
-#else
-
-using ezWindowHandle = void*;
-using ezWindowInternalHandle = ezWindowHandle;
-#  define INVALID_WINDOW_HANDLE_VALUE nullptr
-
-#endif
-
-#ifndef INVALID_INTERNAL_WINDOW_HANDLE_VALUE
-#  define INVALID_INTERNAL_WINDOW_HANDLE_VALUE INVALID_WINDOW_HANDLE_VALUE
-#endif
+#include <WindowDecl_Platform.h>
 
 /// \brief Base class of all window classes that have a client area and a native window handle.
 class EZ_CORE_DLL ezWindowBase
@@ -242,17 +129,17 @@ struct EZ_CORE_DLL ezWindowCreationDesc
 /// Will handle basic message looping. Notable events can be listened to by overriding the corresponding callbacks.
 /// You should call ProcessWindowMessages every frame to keep the window responsive.
 /// Input messages will not be forwarded automatically. You can do so by overriding the OnWindowMessage function.
-class EZ_CORE_DLL ezWindow : public ezWindowBase
+class EZ_CORE_DLL ezWindowPlatformShared : public ezWindowBase
 {
 public:
   /// \brief Creates empty window instance with standard settings
   ///
   /// You need to call Initialize to actually create a window.
   /// \see ezWindow::Initialize
-  ezWindow();
+  ezWindowPlatformShared();
 
   /// \brief Destroys the window if not already done.
-  virtual ~ezWindow();
+  virtual ~ezWindowPlatformShared();
 
   /// \brief Returns the currently active description struct.
   inline const ezWindowCreationDesc& GetCreationDescription() const { return m_CreationDescription; }
@@ -339,18 +226,6 @@ public:
   /// \brief Called when the close button of the window is clicked. Does nothing by default.
   virtual void OnClickClose() {}
 
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS)
-  /// \brief Called on any window message.
-  ///
-  /// You can use this function for example to dispatch the message to another system.
-  ///
-  /// \remarks
-  ///   Will be called <i>after</i> the On[...] callbacks!
-  ///
-  /// \see OnResizeMessage
-  virtual void OnWindowMessage(ezMinWindows::HWND hWnd, ezMinWindows::UINT msg, ezMinWindows::WPARAM wparam, ezMinWindows::LPARAM lparam);
-#endif
-
   /// \brief Returns the input device that is attached to this window and typically provides mouse / keyboard input.
   ezStandardInputDevice* GetInputDevice() const { return m_pInputDevice.Borrow(); }
 
@@ -372,20 +247,10 @@ private:
 
   mutable ezWindowInternalHandle m_hWindowHandle = ezWindowInternalHandle();
 
-#if EZ_ENABLED(EZ_SUPPORTS_GLFW)
-  static void IconifyCallback(GLFWwindow* window, int iconified);
-  static void SizeCallback(GLFWwindow* window, int width, int height);
-  static void PositionCallback(GLFWwindow* window, int xpos, int ypos);
-  static void CloseCallback(GLFWwindow* window);
-  static void FocusCallback(GLFWwindow* window, int focused);
-  static void KeyCallback(GLFWwindow* window, int key, int scancode, int action, int mods);
-  static void CharacterCallback(GLFWwindow* window, unsigned int codepoint);
-  static void CursorPositionCallback(GLFWwindow* window, double xpos, double ypos);
-  static void MouseButtonCallback(GLFWwindow* window, int button, int action, int mods);
-  static void ScrollCallback(GLFWwindow* window, double xoffset, double yoffset);
-#endif
-
   /// increased every time an ezWindow is created, to be able to get a free window index easily
   static ezUInt8 s_uiNextUnusedWindowNumber;
   ezAtomicInteger32 m_iReferenceCount = 0;
 };
+
+// include the platform specific implementation
+#include <Window_Platform.h>
